rename useEffectHmr file
