+++ generator/tools/templates/react/src/hooks/useEffectHmr.tsx
